Moved out common code to util function. Moved test patch functions to utils (Not mapping one for circular imports)
