+++ management_layer/integration.py
@@ -5,7 +5,7 @@ from management_layer.api.stubs import AbstractStubClass
 from management_layer import transformations, mappings
 from management_layer.permission import utils
 from management_layer.permission.decorator import require_permissions, requester_has_role
-from management_layer.utils import client_exception_handler
+from management_layer.utils import client_exception_handler, transform_users_with_roles
 
 TOTAL_COUNT_HEADER = "X-Total-Count"
 CLIENT_TOTAL_COUNT_HEADER = "X-Total-Count"
@@ -533,35 +533,8 @@ class Implementation(AbstractStubClass):
         """
         with client_exception_handler():
             # Get access control response.
-            users_with_roles = await request.app["operational_api"].get_users_with_roles_for_domain(domain_id)
-            users_with_roles = [obj.to_dict() for obj in users_with_roles]
-
-            # Get all user_ids to retrieve.
-            user_ids = [
-                obj["user_id"] for obj in users_with_roles
-            ]
-            if user_ids:
-                # Get all the user names of the user IDs found.
-                users = await request.app[
-                    "authentication_service_api"].user_list(user_ids=user_ids, **kwargs)
-
-                if users:
-                    transform = transformations.USER
-                    users = [transform.apply(user.to_dict()) for user in users]
-                    user_mapping = {
-                        user["id"]: user["username"] for user in users
-                    }
-                    users_with_roles = [
-                        {
-                            "id": user_with_roles["user_id"],
-                            "username": user_mapping[user_with_roles["user_id"]],
-                            "roles": [
-                                mappings.Mappings.role_label_for(role_id)
-                                for role_id in user_with_roles["role_ids"]
-                            ]
-                        } for user_with_roles in users_with_roles
-                    ]
-                return users_with_roles
+            response = await request.app["operational_api"].get_users_with_roles_for_domain(domain_id)
+        return await transform_users_with_roles(request, response, **kwargs)
 
     # get_users_with_roles_for_site -- Synchronisation point for meld
     @staticmethod
@@ -579,35 +552,8 @@ class Implementation(AbstractStubClass):
         """
         with client_exception_handler():
             # Get access control response.
-            users_with_roles = await request.app["operational_api"].get_users_with_roles_for_site(site_id)
-            users_with_roles = [obj.to_dict() for obj in users_with_roles]
-
-            # Get all user_ids to retrieve.
-            user_ids = [
-                obj["user_id"] for obj in users_with_roles if obj["user_id"]
-            ]
-            if user_ids:
-                # Get all the user names of the user IDs found.
-                users = await request.app[
-                    "authentication_service_api"].user_list(user_ids=user_ids, **kwargs)
-
-                if users:
-                    transform = transformations.USER
-                    users = [transform.apply(user.to_dict()) for user in users]
-                    user_mapping = {
-                        user["id"]: user["username"] for user in users
-                    }
-                    users_with_roles = [
-                        {
-                            "id": user_with_roles["user_id"],
-                            "username": user_mapping[user_with_roles["user_id"]],
-                            "roles": [
-                                mappings.Mappings.role_label_for(role_id)
-                                for role_id in user_with_roles["role_ids"]
-                            ]
-                        } for user_with_roles in users_with_roles
-                    ]
-                return users_with_roles
+            response = await request.app["operational_api"].get_users_with_roles_for_site(site_id)
+        return await transform_users_with_roles(request, response, **kwargs)
 
     # get_domain_roles -- Synchronisation point for meld
     @staticmethod
